Set a consistent heat/cool mode when all devices are in dead zone
When all devices are in the dead zone, the average low, high and current temperature is taken. If the current temperature is closer to high, all heatpumps are set to cooling. If the average is closer to low, all heatpumps are set to heating.

Given that they're all in the deadzone, no heating or cooling will actually take place, but fans will continue to run.