Reliability improvements, better fallback mode state for zone consistency and saving HEAT+COOL mode to flash
For the case where all zones are within their target setpoints, we'll revert to setting HEAT/COOL to the last known state instead of calculating an overall average. The reasoning here is the rooms that most recently required heating/cooling will probably require them again before the other rooms that may already be close to their setpoint.

Instead of waiting for temperature changes, we'll now ensure the heat pump is set to the desired mode on every sync call. This should work around a bug where the cooling setpoint temperature may accidentally be set while the unit is in heat mode.

Previously managed mode (HEAT/COOL) would be automatically disabled after a power cycle. This setting is now persisted to disk, although I do need to figure out a way to automatically disable this in the event that a home assistant connection could not be made.